refactor : 테스트 계정 | 테스트 계정 생성 로직 변경
