+++ Backend/sdutyplus/build.gradle
@@ -79,6 +79,9 @@ dependencies {
 	//Jasypt
 	implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.4'
 	implementation 'org.bouncycastle:bcprov-jdk15on:1.61'
+
+	//Apache Commons Lang for Random String
+	implementation 'org.apache.commons:commons-lang3:3.12.0'
 }
 
 //querydsl start
