+++ Backend/sdutyplus/src/main/resources/application.properties
@@ -5,9 +5,9 @@ jasypt.encryptor.bean=jasyptStringEncryptor
 
 #DB Setting
 #develop - local
-#spring.datasource.url=ENC(cCCs4sLD03dDgwHpyL/RCWpydbevC7onaizDd3JsaxfanVeVjUWqy4y2kQa/JSktAn6AOm2IjOs=)
+spring.datasource.url=ENC(cCCs4sLD03dDgwHpyL/RCWpydbevC7onaizDd3JsaxfanVeVjUWqy4y2kQa/JSktAn6AOm2IjOs=)
 #develop - server
-spring.datasource.url=ENC(dVZ/1Rcwlv3ZnDmQ9zDilDYW/tVJRtirW5QPX9qRvN3uk4sbF0jxxC0iAhG5SG/QdY/3ZnehkLVeVJIfvZznQw==)
+#spring.datasource.url=ENC(dVZ/1Rcwlv3ZnDmQ9zDilDYW/tVJRtirW5QPX9qRvN3uk4sbF0jxxC0iAhG5SG/QdY/3ZnehkLVeVJIfvZznQw==)
 
 #dev - local
 #spring.datasource.url-ENC(exyCjCGa7x5VQGwJhHkgPgYzHMdUJkQe4KmM/IzE1RC81yoB9EiI2S1nW2kH/8tI)
