+++ Backend/sdutyplus/src/test/java/com/d205/sdutyplus/IntegrationTest.java
@@ -1,0 +1,15 @@
+package com.d205.sdutyplus;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+@Transactional
+@SpringBootTest
+public class IntegrationTest {
+
+    @PersistenceContext
+    protected EntityManager em;
+}
